mv eventsFileHeader and eventsFileVersion consts to the inter/event_serializer
